Moved GoogleApiClient to BaseActivity
